Datos atmosféricos

Datos atmosféricos

atmosfera

Format

Un data.frame con 41472 filas y 11 columnas

latitud

Ubicación de la medida. Grilla espacial de 24 por 24 desde longitud 113.8O a 56.2O y desde latitud 36.2N a 21.2S.

longitud

Ubicación de la medida. Grilla espacial de 24 por 24 desde longitud 113.8O a 56.2O y desde latitud 36.2N a 21.2S.

anio

72 puntos en el tiempo. Una vez al mes desde enero de 1995 hasta diciembre de 2000.

mes

72 puntos en el tiempo. Una vez al mes desde enero de 1995 hasta diciembre de 2000.

temp_superficie

Temperatura media mensual basada en la energía emitida desde la Tierra en condiciones de cielo despejado (en K).

temperatura

Temperatura media mensual del aire cerca de la superficie de la Tierra (en K).

presion

Presión media mensual en una ubicación determinada de la superficie de la Tierra (en mb).

ozono

Cantidad media mensual de ozono en la columna atmosférica (en Dobsons).

nube_baja

Porcentaje de la media mensual de cielo cubierto por nubes con una presión superior mayor a 680 mb o aproximadamente menor que 3.24 km.

nube_media

Porcentaje de la media mensual de cielo cubierto por nubes con una presión superior entre 440-680 mb o aproximadamente menor que 3.24 a 6.5 km.

nube_alta

Porcentaje de la media mensual de cielo cubierto por nubes con una presión superior menor o igual a 440 mb o aproximadamente mayor que 6.5 km.