Tabla de estadísticas de bateo

Estadísticas de bateadores

Format

Un data frame con 105.861 filas y 22 columnas

id_jugador

ID del jugador

id_anio

Año

orden_equipos

Orden en que el jugador se movió entre equipos dentro de la misma temporada

id_equipo

ID del equipo (factor)

id_liga

ID de la liga (factor con niveles AA, AL, FL, NL, PL, UA)

juegos

Número de juegos en los que participó oficialmente el jugador

al_bate

Turnos al bate

carreras

Carreras anotadas

golpes

Cantidad de veces en la que el jugador alcanzó una base debido a un golpe, sin error de la defensa

dobles

Hits en los que el bateador llegó a segunda base

triples

Golpes en los que el bateador llegó a tercera base

cuadrangulares

Cuadrangulares - jonrones

carreras_empujadas

Carreras empujadas

bases_robadas

Bases robadas

atrapado_robando

Atrapado (out) en intento de robo

base_bolas

Base por bolas

ponches

Ponches

base_intencional

Base por bolas intencional

golpeado

Golpeado por lanzamiento

toque_sacrificio

Toque de sacrificio

elavado_sacrificio

Elevado de sacrificio

doble_matanza

Batazos para doble matanza (doble play)