Precio de 50.000 diamantes

Un conjunto de datos que contiene los precios de casi 54.000 diamantes.

diamantes

Format

Un data.frame con 53.940 filas y 10 columnas

precio

Precio en dólares americanos (\$326-\$18,823)

quilate

Peso del diamante (0.2-5.01)

corte

Calidad del corte (Regular, Bueno, Muy bueno, Premium, Ideal)

color

Color del diamante, de J (peor) a D (mejor)

claridad

Medida de que tan claro es el diamante (I1 (peor), SI1, SI2, VS1, VS2, VVS1, VVS2, IF (mejor))

profundidad

Porcentaje de la profundidad total = z / mean(x, y) = 2 * z / (x + y) (43-79)

tabla

Ancho de la parte superior del diamante con relación a su punto más ancho (43-95)

x

Largo en milimetros

y

Ancho en milimetros

z

Profundidad en milimetros