Tabla de estadísticas de lanzadores

Estadísticas de lanzadores

Format

Un data frame con 46.699 filas y 30 columnas

id_jugador

ID del jugador

id_anio

Año

orden_equipos

Orden en que el jugador se movió entre equipos dentro de la misma temporada

id_equipo

ID del equipo (factor)

id_liga

ID de la liga (factor con niveles AA, AL, FL, NL, PL, UA)

juegos_ganados

Juegos jugados ganados

juegos_perdidos

Juegos jugados perdidos

juegos_jugados

Juegos jugados

juegos_iniciados

Juegos jugados iniciados

juegos_completados

Número de juegos completos (9 innings lanzados) que lanzó el jugador

blanqueadas

Número de blanqueos (juegos completos sin permitir carreras) que lanzó el jugador

juegos_salvados

Juegos salvados

IPouts

Outs propinados al equipo contrario (Igual a innings lanzados x 3)

hits

Hits permitidos del oponente

carreras_ganadas

Carreras limpias recibidas

cuadrangulares

Cuadrangulares recibidos

BB

Base por bolas cedidas al oponente

ponches

Ponches propinados al oponente

promedio_bateo_rival

Promedio de bateo del rival

promedio_carreras_ganadas

Promedio de carreras limpias permitidas (normalizada a 9 innings lanzados)

IBB

Base por bolas intencionales cedidas al oponente

lanzamientos_desviados

Lanzamientos desviados lanzados por el lanzador

HBP

Bateadores golpeador por el lanzador

BK

Balks (movimiento ilegal del cuerpo realizado por el lanzador)

BFP

Bateadores a los que el lanzador se ha enfrentado

juegos_finalizados

Juegos en los que el lanzador finalizó el juego

carreras

Carreras recebidas (sucias y limpias)

sacrificios_golpeados

Toques de sacrificio que el oponente le hizo al lanzador

vuelos_sacrificio

Elevado (fly) de sacrificio que el oponente le hizo al lanzador

doble_matanza

Doble matanza inducidas por el lanzador