Tabla de personas

Nombres de jugadores, fecha de nacimiento e información bibliográfica. Este archivo se usa para obtener detalles acerca de los jugadores listados en bateadores, lanzadores y otros archivos en que los jugadores está identificador por id_jugador.

Format

Un data frame con 19.617 filas y 26 columnas

id_jugador

ID del jugador

anio_nacimiento

Año de nacimiento

mes_nacimiento

Mes de nacimiento

dia_nacimiento

Día de nacimiento

pais_nacimiento

País de nacimiento

estado_nacimiento

Estado de nacimiento

ciudad_nacimiento

Ciudad de nacimiento

anio_fallecimiento

Año de fallecimiento

mes_fallecimiento

Mes de fallecimiento

dia_fallecimiento

Día de fallecimiento

pais_fallecimiento

País de fallecimiento

estado_fallecimiento

Estado de fallecimiento

ciudad_fallecimiento

Ciudad de fallecimiento

nombre

Nombre

apellido

Apellido

nombre_completo

Nombre y apellido

peso

Peso en libras

estatura

Estatura en pulgadas

bateador

Diestro (D) o zurdo (Z)

lanzador

Diestro (D) o zurdo (Z)

fecha_primer_juego

Fecha primer juego en primera división

fecha_ultimo_juego

Fecha último juego en primera división

id_retrosheet

Código de jugador usado en Retrosheet (http://www.retrosheet.org/)

id_brw

Código de jugador usado en Baseball Reference website (http://www.baseball-reference.com/)

fecha_fallecimiento

Fecha de fallecimiento en formato yyyy-mm-dd

fecha_nacimiento

Fecha de nacimiento en formato yyyy-mm-dd