Datos de vuelos

Datos temporales de todos los vuelos que despegaron de Nueva York (JFK, LGA o EWR) durante 2013.

vuelos

Format

Un data.frame con 336.776 filas y 19 columnas

anio

año de la fecha de salida

mes

mes de la fecha de salida

dia

día de la fecha de salida

horario_salida

horario efectivo de salida del vuelo (formato HHMM o HMM), hora local

salida_programada

horario programado para la salida (formato HHMM o HMM), hora local

atraso_salida

atraso de la salida en minutos. Valores negativos indican salida adelantada

horario_llegada

horario efectivo de llegada del vuelo (formato HHMM o HMM), hora local

llegada_programada

horario programado para la llegada (formato HHMM o HMM), hora local

atraso_llegada

atraso de la llegada en minutos. Valores negativos indican llegada adelantada

aerolinea

abreviación de dos letras de la aerolínea. Ver `aerolineas` para obtener el nombre

vuelo

número de vuelo

codigo_cola

código de cola del avión

origen

origen del vuelo. Ver `aeropuertos` para metadatos adicionales

destino

destino del vuelo. Ver `aeropuertos` para metadatos adicionales

tiempo_vuelo

cantidad de tiempo en aire, en minutos

distancia

distancia entre aeropuertos, en millas

hora

hora del horario programado para la salida

minuto

minutos del horario programado para la salida

fecha_hora

fecha y horario programados del vuelo en formato POSIXct